Move function to importing folder
