+++ satay/transposonmapping/python_modules/__init__.py
@@ -1,10 +1,9 @@
-from .verify_data_files import verify_data_files
 from .loading_files import access_files
 from .chromosome_and_gene_positions import chromosomename_roman_to_arabic, gene_position
 from .gene_names import gene_aliases
 from .samflag import samflags
 from .get_chromosome_properties import get_chromosome_names, get_sequence_length, get_chromosome_reads
-
+from .get_reads import get_reads
 
 __all__ = [
     access_files, 
@@ -13,7 +12,7 @@ __all__ = [
     gene_position, 
     get_chromosome_names,
     get_chromosome_reads,
+    get_reads,
     get_sequence_length,
     samflags, 
-    verify_data_files
     ]
